task: moved normal menu inside the dropdown menu
